Add operators
